refact: refact code
