+++ rs-commerce/src/views/registration/Registration.tsx
@@ -2,7 +2,7 @@ import RegistrationForm from './components/RegistrationForm/RegistrationForm';
 import TitleRegistration from './components/TitleRegistration/TitleRegistration';
 import classes from './style.module.css';
 
-function Registration() {
+function Registration(): JSX.Element {
   return (
     <div className={classes['registration']}>
       <div className={classes['registration__wrapper']}>
