+++ rs-commerce/src/views/registration/components/RegistrationForm/components/InputRegistration/CountryInput.tsx
@@ -1,20 +1,17 @@
-import { AddressType } from '../AddressRegistration/components/AddressComponents/components/AddresTitleComponent/AddresTitleComponent';
+import { AddressType, TypeCountry } from 'types/registrationTypes';
 import classes from './styles.module.css';
 import { registrationParamsObj } from './utils/checkFields';
-
-type TypeCountry = 'shippingCountry' | 'billingCountry';
 
 const countryArr = ['USA', 'Belarus'];
 
 function CountryInput(props: AddressType): JSX.Element {
   const { typeComponent } = props;
   function getValue(e: React.ChangeEvent<HTMLInputElement>) {
-    console.log(e.target.value);
     const field = `${typeComponent}Country` as TypeCountry;
     registrationParamsObj[field] = e.target.value;
   }
 
-  function clearInput(e: React.FocusEvent<HTMLInputElement, Element>) {
+  function clearInput(e: React.FocusEvent<HTMLInputElement, Element>): void {
     e.target.value = '';
   }
 
@@ -28,6 +25,7 @@ function CountryInput(props: AddressType): JSX.Element {
           onChange={(e) => getValue(e)}
           onFocus={(e) => clearInput(e)}
           list="country"
+          placeholder="Choose your country"
         />
         <datalist id="country">
           {countryArr.map((item) => (
