+++ rs-commerce/src/views/registration/components/TitleRegistration/TitleRegistration.tsx
@@ -1,11 +1,11 @@
 import classes from './style.module.css';
 
-function TitleRegistration() {
+function TitleRegistration(): JSX.Element {
   return (
     <div className={classes['registration__title-block']}>
       <h1 className={classes['registration__title']}>Registration</h1>
       <div className={classes['registration__login-block']}>
-        <p className={classes['registration__login-content']}>Have you already had an account?</p>
+        <p className={classes['registration__login-content']}>I have an account?</p>
         <p className={classes['registration__link']}>Login</p>
       </div>
     </div>
